requirements.txt
appdirs==1.4.3
asgiref==3.2.10
bcrypt==3.1.7
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.0
django-cors-headers==3.5.0
django-seed==0.2.2
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Faker==4.17.1
html5lib==1.0.1
idna==2.10
inflection==0.5.0
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pep517==0.8.2
Pillow==7.1.2
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
